Merge lexer and parser back together
Earlier in the git history there are commits which removed the
lexer-related files so as to separate lexer and parser into separate
projects. However, these projects share a lot of source code so it
makes much more sense to keep them together. This commit reverses
the earlier commits that removed the lexer.